Common::Configuration::DTO

Module ConfigurationOverridesDTO

ConfigurationOverridesDTO

Defines the Data Transfer Object for specifying configuration overrides when retrieving or inspecting values.

Structs

ConfigurationOverridesDTO

A serializable struct for specifying overrides when retrieving or inspecting configuration values.