Common::DTO

Module WorkSpaceEditDTO

WorkSpaceEditDTO

Defines the Data Transfer Object for a workspace edit, which is a collection of changes to be applied across multiple files in the workspace.

Structs

WorkSpaceEditDTO

A serializable struct representing a workspace edit, which is a batch of changes that can include text edits and file operations (create, delete, rename). This is analogous to vscode.WorkspaceEdit.