Common::Environment::Requires

Trait Requires

pub trait Requires<Capability: ?Sized>: Environment {
    // Required method
    fn Require(&self) -> Arc<Capability>;
}

A trait that enables an environment (Self) to provide a specific capability (Capability).

This is the central mechanism for dependency injection. An ActionEffect can be generic over an environment TEnvironment as long as it has the bound TEnvironment: Requires<MyCapability>. The effect can then call Environment.Require() to receive a shared instance of the service it needs to perform its operation.

The Capability is typically a trait object, such as dyn FileSystemReader.

Required Methods

fn Require(&self) -> Arc<Capability>

Returns the required capability from the environment, wrapped in an Arc for safe, shared ownership.