Common::Error::CommonError

Enum CommonError

pub enum CommonError {

    FileSystemIO {
        Path: PathBuf,
        Description: String,
    },
    FileSystemNotFound(PathBuf),
    FileSystemPermissionDenied {
        Path: PathBuf,
        Reason: String,
    },
    FileSystemFileExists(PathBuf),
    FileSystemNotADirectory(PathBuf),
    FileSystemIsADirectory(PathBuf),
    ConfigurationUpdate {
        Key: String,
        Description: String,
    },
    ConfigurationLoad {
        Description: String,
    },
    InvalidArgument {
        ArgumentName: String,
        Reason: String,
    },
    NotImplemented {
        FeatureName: String,
    },
    StateLockPoisoned {
        Context: String,
    },
    IPCError {
        Description: String,
    },
    CommandExecution {
        CommandIdentifier: String,
        Reason: String,
    },
    CommandNotFound {
        Identifier: String,
    },
    ProviderRegistration {
        ProviderType: String,
        Reason: String,
    },
    ProviderInvocation {
        ProviderIdentifier: String,
        Reason: String,
    },
    TreeViewProviderNotFound {
        ViewIdentifier: String,
    },
    ExternalServiceError {
        ServiceName: String,
        Description: String,
    },
    SecretsAccess {
        Key: String,
        Reason: String,
    },
    UserInterfaceInteraction {
        Reason: String,
    },
    SerializationError {
        Description: String,
    },
    Unknown {
        Description: String,
    },

}

A common error enum for the application, encompassing all major categories of failures that can occur during the execution of effects or other

Variants

FileSystemIO

Fields

Path: PathBuf

Description: String

FileSystemNotFound(PathBuf)

FileSystemPermissionDenied

Fields

Path: PathBuf

Reason: String

FileSystemFileExists(PathBuf)

FileSystemNotADirectory(PathBuf)

FileSystemIsADirectory(PathBuf)

ConfigurationUpdate

Fields

Key: String

Description: String

ConfigurationLoad

Fields

Description: String

InvalidArgument

Fields

ArgumentName: String

Reason: String

NotImplemented

Fields

FeatureName: String

StateLockPoisoned

Fields

Context: String

IPCError

Fields

Description: String

CommandExecution

Fields

CommandIdentifier: String

Reason: String

CommandNotFound

Fields

Identifier: String

ProviderRegistration

Fields

ProviderType: String

Reason: String

ProviderInvocation

Fields

ProviderIdentifier: String

Reason: String

TreeViewProviderNotFound

Fields

ViewIdentifier: String

ExternalServiceError

Fields

ServiceName: String

Description: String

SecretsAccess

Fields

Key: String

Reason: String

UserInterfaceInteraction

Fields

Reason: String

SerializationError

Fields

Description: String

Unknown

Fields

Description: String

Implementations

impl CommonError

pub fn FromStandardIOError( IOError: Error, Path: PathBuf, OperationContext: &str, ) -> Self

Creates a CommonError from a standard std::io::Error, mapping common I/O error kinds to our specific FileSystem error variants for better contextualization.