Common::LanguageFeature::DTO

Module ProviderType

ProviderType DTO

Defines the enum that identifies each type of language feature provider.

Enums

ProviderType

An enum that provides a unique identifier for each type of language feature. This is used to register and query for specific provider implementations.