Trait common::arithmetic::bounds::Bounded

pub trait Bounded {
    // Required methods
    fn min_value() -> Self;
    fn max_value() -> Self;
}

Numbers which have upper and lower bounds

Required Methods

fn min_value() -> Self

Returns the smallest finite number this type can represent

fn max_value() -> Self

Returns the largest finite number this type can represent