common::ast::node_type

Enum NodeType

pub enum NodeType {

    NoExpression,
    Literal(String),
    Identifier(String),
    Operator(String),
    Type(DataType),
    Constant(String),
    IfStatement,
    ElseStatement,
    ForLoop,
    WhileLoop,
    DoWhileLoop,
    Break,
    Continue,
    Return,
    SwitchStatement,
    Case,
    Default,
    Assignment,
    Initialization,
    FunctionDeclaration,
    StructDeclaration,
    EnumDeclaration,
    ModuleExpression,
    TopLevelExpression,
    BlockExpression,
    Condition,
    Action,
    Variant,
    AssignedValue,
    Field,
    Parameter,
    Variable,
    BinaryExpression,
    UnaryExpression,
    FunctionCall,
    Operand,
    LoopInitializer,
    LoopIncrement,

}

Defines acceptable syntax elements as part of an AST.

Each element A different kind of syntactic construct that can appear in source code, such as constants, identifiers, operators, and control structures. These elements are used to build a tree representation of the codeโs syntactic structure.

Variants

NoExpression

A lack of expression, often used as a placeholder.

Literal(String)

A literal value, stored as a string.

Identifier(String)

An identifier, such as a variable name, stored as a string.

Operator(String)

An operator, such as โ+โ, โ-โ, etc., stored as a string.

Type(DataType)

A data type, linking directly to the DataType enum.

Constant(String)

A constant value, stored as a string.

IfStatement

An if statement.

ElseStatement

An else statement.

ForLoop

A for loop.

WhileLoop

A while loop.

DoWhileLoop

A do-while loop.

Break

A break statement.

Continue

A continue statement.

Return

A return statement.

SwitchStatement

A switch statement.

Case

A case in a switch statement.

Default

The default branch in a switch statement.

Assignment

An assignment of a value to an existing variable.

Initialization

The initialization of a variable.

FunctionDeclaration

A function declaration.

StructDeclaration

A struct declaration.

EnumDeclaration

An enum declaration.

ModuleExpression

A module-level expression, often used for scoping.

TopLevelExpression

A top-level expression in a module or file.

BlockExpression

A block of expressions or statements.

Condition

A condition in control flows.

Action

An action to be taken, typically in control structures.

Variant

A variant in enums or other similar structures.

AssignedValue

A value assigned in statements like initialization and assignment.

Field

A field in a struct or similar data structure.

Parameter

A parameter in function declarations.

Variable

A variable, used in various expressions and statements.

BinaryExpression

A binary expression, involving two operands and an operator.

UnaryExpression

A unary expression, involving one operand and an operator.

FunctionCall

A function call.

Operand

An operand in an expression.

LoopInitializer

The initializer section of a loop.

LoopIncrement

The increment section of a loop.