common::io

Function write_ir_to_file

pub fn write_ir_to_file(
    module: Arc<RwLock<SafeLLVMPointer>>,
    file_name: &str,
) -> Result<(), String>

Writes an LLVM module to a file.

This function serializes an LLVM module, contained inside a SafeLLVMPointer, to a specified file.

Parameters

module - An Arc<RwLock<SafeLLVMPointer>> pointing to the LLVM module to be written.

file_name - The name of the file where the LLVM IR should be saved.

Returns

A Result<(), String> indicating the success or failure of the operation. Returns Ok(()) if the module is successfully written to the file. Returns Err(String) if there are issues obtaining locks, converting paths, creating directories, or in the LLVM API call to write the module.