Trait common::storage::StorageDecodeIsFull

pub trait StorageDecodeIsFull: StorageDecodeLength {
    // Required method
    fn bound() -> usize;

    // Provided method
    fn decode_is_full(key: &[u8]) -> Option<bool> { ... }
}

Required Methods

fn bound() -> usize

Bound on the length of the collection

Provided Methods

fn decode_is_full(key: &[u8]) -> Option<bool>

Decode if the storage value at key is full.

This function assumes that the length is at the beginning of the encoded object and is a Compact<u32>.

Returns None if the storage value does not exist or the decoding failed.