Trait common::LiquiditySource

pub trait LiquiditySource<TargetId, AccountId, AssetId, Amount, Error> {
    // Required methods
    fn can_exchange(
        target_id: &TargetId,
        input_asset_id: &AssetId,
        output_asset_id: &AssetId
    ) -> bool;
    fn quote(
        target_id: &TargetId,
        input_asset_id: &AssetId,
        output_asset_id: &AssetId,
        amount: QuoteAmount<Amount>,
        deduce_fee: bool
    ) -> Result<(SwapOutcome<Amount>, Weight), Error>;
    fn step_quote(
        target_id: &TargetId,
        input_asset_id: &AssetId,
        output_asset_id: &AssetId,
        amount: QuoteAmount<Amount>,
        recommended_samples_count: usize,
        deduce_fee: bool
    ) -> Result<(VecDeque<SwapChunk<Amount>>, Weight), Error>;
    fn exchange(
        sender: &AccountId,
        receiver: &AccountId,
        target_id: &TargetId,
        input_asset_id: &AssetId,
        output_asset_id: &AssetId,
        swap_amount: SwapAmount<Amount>
    ) -> Result<(SwapOutcome<Amount>, Weight), Error>;
    fn check_rewards(
        target_id: &TargetId,
        input_asset_id: &AssetId,
        output_asset_id: &AssetId,
        input_amount: Amount,
        output_amount: Amount
    ) -> Result<(Vec<(Amount, AssetId, RewardReason)>, Weight), Error>;
    fn quote_without_impact(
        target_id: &TargetId,
        input_asset_id: &AssetId,
        output_asset_id: &AssetId,
        amount: QuoteAmount<Amount>,
        deduce_fee: bool
    ) -> Result<SwapOutcome<Amount>, Error>;
    fn quote_weight() -> Weight;
    fn step_quote_weight(samples_count: usize) -> Weight;
    fn exchange_weight() -> Weight;
    fn check_rewards_weight() -> Weight;
}

Indicates that particular object can be used to perform exchanges.

Required Methods

fn can_exchange( target_id: &TargetId, input_asset_id: &AssetId, output_asset_id: &AssetId ) -> bool

Check if liquidity source provides an exchange from given input asset to output asset.

fn quote( target_id: &TargetId, input_asset_id: &AssetId, output_asset_id: &AssetId, amount: QuoteAmount<Amount>, deduce_fee: bool ) -> Result<(SwapOutcome<Amount>, Weight), Error>

Get spot price of tokens based on desired amount.

fn step_quote( target_id: &TargetId, input_asset_id: &AssetId, output_asset_id: &AssetId, amount: QuoteAmount<Amount>, recommended_samples_count: usize, deduce_fee: bool ) -> Result<(VecDeque<SwapChunk<Amount>>, Weight), Error>

Get the input/output liquidity divided into steps based on the desired amount. The count of steps may differ from the recommended_samples_count

fn exchange( sender: &AccountId, receiver: &AccountId, target_id: &TargetId, input_asset_id: &AssetId, output_asset_id: &AssetId, swap_amount: SwapAmount<Amount> ) -> Result<(SwapOutcome<Amount>, Weight), Error>

Perform exchange based on desired amount.

fn check_rewards( target_id: &TargetId, input_asset_id: &AssetId, output_asset_id: &AssetId, input_amount: Amount, output_amount: Amount ) -> Result<(Vec<(Amount, AssetId, RewardReason)>, Weight), Error>

Get rewards that are given for perfoming given exchange.

fn quote_without_impact( target_id: &TargetId, input_asset_id: &AssetId, output_asset_id: &AssetId, amount: QuoteAmount<Amount>, deduce_fee: bool ) -> Result<SwapOutcome<Amount>, Error>

Get spot price of tokens based on desired amount, ignoring non-linearity of underlying liquidity source.

fn quote_weight() -> Weight

Get weight of quote

fn step_quote_weight(samples_count: usize) -> Weight

Get weight of step quote

fn exchange_weight() -> Weight

Get weight of exchange

fn check_rewards_weight() -> Weight

Get weight of exchange