Crate common

Modules

crc

logging

profiling

signals

Macros

mca_debug

Get global debug level

mca_dformat

mca_dformat! - diagnostic formatted message with variable args Formatting string is template for args

mca_dprintf

mca_dprintf! - diagnostic print filtered by level Debug level threshold is set with environment variable MCA_DEBUG Formatting string is template for args

Constants

MCA_DOMAIN_INVALID

MCA_FALSE

MCA_HANDLE_MASK

MCA_INFINITE

MCA_MAX_CPUS

MCA_MAX_DOMAINS

MCA_MAX_NODES

MCA_MAX_PROCESSES

MCA_MAX_REQUESTS

MCA_NODE_INVALID

MCA_NULL

MCA_RETURN_VALUE_INVALID

MCA_TRUE

Statics

MCA_DEBUG

MCA_DEBUG_INITIALIZED

Type Definitions

McaBoolean

McaDomain

McaInt

McaInt8

McaInt16

McaInt32

McaInt64

McaInt128

McaNode

McaStatus

McaTimeout

McaUint

McaUint8

McaUint16

McaUint32

McaUint64

McaUint128