Crate common

Common

This crate contains various utility functions share across the repository:

cache contains a generic static container implemented used for caching.

stream contains functionality needed for sending a File Descriptor across a Unix Socket.

singleton contains an implementation of a Reentrant Synchronization Singleton for protecting critical paths from multiple threads modifying the global state.

Modules

cache

This file contains an interface to a static DashMap, such that the stored values can be extracted with static lifetimes, and without a Ref Guard.

singleton

This file implements a Reentrant Synchronization Singleton, which is used to guard a critical path such that only a single thread may execute it at a time. The general use is:

stream

This file contains various utilities dealing with Sockets and Streams.