Crate common

the common crate defines all the common structs, functions and enums

It provides tooling such as logging, environment variable config and so on

Structs

Book

A Book represents a level 2 order book data structure

BookSnapshot

Orderbook snapshot message

Context

Worker Context

Error

HttpClient

An Http client for Rest APIs

InconsistentBook

L2

One side of a Level 2 Order book

L2Iterator

NewBookSnapshot

Gateway command

WsConsumer

Websocket consumer

WsInfo

WsPayload

Websocket payload

Enums

ErrorKind

InnerMessage

Internal message enum

Side

Orderbook side

Traits

Request

An HTTP Request

Functions

bid_ask_spread

Calculate ask-bid spread

create_config

init_logging

Initialise slog

wrap_result

Type Definitions

CfgBuilder

WorkerContext

A context for a courotine worker