Crate common

Re-exports

pub use fixnum;

pub use paste;

pub use macros::*;

pub use utils::*;

Modules

eth

macros

migrations

prelude

serialization

utils

weights

Macros

assert_approx_eq

balance

dbg

fixed

fixed_const

fixed_wrapper

generate_storage_instance

location_stamp

our_include

our_include_bytes

storage_remove_all

vec_push

Structs

AssetId32

This is wrapped structure, this is like H256 or Р512, extra PhantomData is added for typing reasons.

AssetName

AssetSymbol

BalanceWrapper

Wrapper type which extends Balance serialization, used for json in RPC’s.

ContentSource

DEXInfo

Information about state of particular DEX.

Description

LiquiditySourceId

Identification of liquidity source.

PswapRemintInfo

Rate

Information about received oracle symbol (price and last update time)

SymbolName

TradingPair

Trading pair data.

Enums

AssetIdExtraAssetRecordArg

DEXId

DEX identifier.

FilterMode

InvokeRPCError

Common error which can arise while invoking particular RPC call in runtime.

LiquiditySourceType

Enumaration of all available liquidity sources.

ManagementMode

Oracle

List of available oracles

PredefinedAssetId

Asset identifier.

PriceVariant

RewardReason

Reason for particular reward during swap.

TechAccountId

Enum encoding of technical account id, pure and wrapped records. Enum record WrappedRepr is wrapped represention of Pure variant of enum, this is useful then representation is known but backward mapping is not known.

TechAssetId

TechPurpose

Code of purpose for technical account.

Constants

ASSET_CONTENT_SOURCE_MAX_LENGTH

Max length of asset content source. The same value as IE URL length. It should enough for any URI / IPFS address (CID)

ASSET_DESCRIPTION_MAX_LENGTH

Max length of asset description, it should be enough to describe everything the user wants

CERES_ASSET_ID

DAI

DEFAULT_BALANCE_PRECISION

DEMETER_ASSET_ID

DOT

ETH

FIXED_PRECISION

HERMES_ASSET_ID

KSM

PSWAP

TBCD

TECH_ACCOUNT_MAGIC_PREFIX

This data is used as prefix in AccountId32, if it is representative for TechAccId encode twox hash (128 + 128 = 256 bit of AccountId32 for example).

USDT

VAL

XOR

XST

XSTUSD

Traits

Config

Common DEX trait. Used for DEX-related pallets.

DataFeed

DataFeed trait indicates that particular object could be used for querying oracle data.

DemeterFarmingPallet

EnsureDEXManager

Check on origin that it is a DEX owner.

EnsureTradingPairExists

FromGenericPair

Abstract trait to get data type from generic pair name and data.

GetLstIdAndTradingPairFromTechAsset

GetMarketInfo

Trait for bounding liquidity proxy associated type representing primary market in TBC.

GetPoolReserves

Trait for bounding liquidity proxy associated type representing secondary market.

GetTechAssetWithLstTag

IsRepresentable

IsRepresentation

IsValid

LiquidityProxyTrait

Indicates that particular object can be used to perform exchanges with aggregation capability.

LiquidityRegistry

LiquiditySource

Indicates that particular object can be used to perform exchanges.

OnNewSymbolsRelayed

Hook-like trait for oracles to capture newly relayed symbols.

OnPoolCreated

OnPoolReservesChanged

OnPswapBurned

General trait for passing pswap amount burned information to required pallets.

OnValBurned

General trait for passing on the amount of burned VAL.

PoolXykPallet

PriceToolsPallet

PureOrWrapped

SwapAction

Definition of a pending atomic swap action. It contains the following three phrases:

SwapRulesValidation

ToFeeAccount

ToMarkerAsset

ToTechUnitFromDEXAndAsset

ToTechUnitFromDEXAndTradingPair

VestedRewardsPallet

Trait to abstract interface of VestedRewards pallet, in order for pallets with rewards sources avoid having dependency issues.

WrappedRepr

Functions

comm_merkle_op

Commutative merkle operation, is crypto safe, defined as hash(a,b) xor hash(b,a).

convert_block_number_to_timestamp

Converts block_number to timestamp

hash

hash_to_u128_pair

sort_with_hash_key

Sorting of keys and values by key with hash_key, useful for crypto sorting with commutative merkle operator.

with_transaction

Similar to #[transactional]

Type Definitions

AccountIdOf

Amount

Asset

Basic type representing asset.

AssetId32Code

This code is H256 like.

Balance

BalancePrecision

BasisPoints

Type definition representing financial basis points (1bp is 0.01%)

DexIdOf

Fixed

Basic type representing assets quantity.

FixedInner

Price