Crate common

Re-exports

pub use fixnum;

pub use paste;

pub use fixed_wrapper_u256::*;

pub use macros::*;

pub use utils::*;

Modules

alt

arithmetic

cache_storage

eth

fixed

fixed_wrapper_u256

macros

migrations

permissions

prelude

serialization

storage

utils

weights

Macros

assert_approx_eq

Assertion if two numbers left and right are equal up to some tolerance.

assert_approx_eq_abs

Assertion that two values are approximately equal up to some absolute tolerance (constant value)

assert_approx_eq_rel

Assertion that two values are approximately equal up to some relative tolerance (percentage of their magnitude a.abs() + b.abs())

balance

dbg

fixed

fixed_const

fixed_u256_float

fixed_u256_int

fixed_wrapper

fixed_wrapper_u256_float

fixed_wrapper_u256_int

generate_storage_instance

location_stamp

our_include

our_include_bytes

storage_remove_all

vec_push

Structs

AssetId32

This is wrapped structure, this is like H256 or Р512, extra PhantomData is added for typing reasons.

AssetInfo

Presents information about an asset.

AssetName

AssetSymbol

BalanceWrapper

Wrapper type which extends Balance serialization, used for json in RPC’s.

BoundedString

ContentSource

CrowdloanTag

DEXInfo

Information about state of particular DEX.

Description

LiquiditySourceId

Identification of liquidity source.

LiquiditySourceTypeIter

An iterator over the variants of LiquiditySourceType

OrderBookId

PswapRemintInfo

Rate

Information about received oracle symbol (price and last update time)

SymbolName

TradingPair

Trading pair data.

Enums

AssetIdExtraAssetRecordArg

AssetType

Enumeration of all supported asset types.

DEXId

DEX identifier.

FilterMode

InvokeRPCError

Common error which can arise while invoking particular RPC call in runtime.

LiquiditySourceType

Enumaration of all available liquidity sources.

ManagementMode

Oracle

List of available oracles

PredefinedAssetId

Asset identifier.

PredefinedSbtAssetId

Predefined SBT asset identifier.

PriceVariant

RewardReason

Reason for particular reward during swap.

TechAccountId

Enum encoding of technical account id, pure and wrapped records. Enum record WrappedRepr is wrapped represention of Pure variant of enum, this is useful then representation is known but backward mapping is not known.

TechAssetId

TechPurpose

Code of purpose for technical account.

Constants

APOLLO_ASSET_ID

ASSET_CONTENT_SOURCE_MAX_LENGTH

Max length of asset content source. The same value as IE URL length. It should enough for any URI / IPFS address (CID)

ASSET_DESCRIPTION_MAX_LENGTH

Max length of asset description, it should be enough to describe everything the user wants

ASSET_ID_PREFIX_KENSETSU_PEGGED_TO_ORACLE

Kensetsu asset ids pegged to oracle start with 0x05…

ASSET_ID_PREFIX_KENSETSU_PEGGED_TO_SORA

Kensetsu asset ids pegged to sora assets start with 0x04…

ASSET_ID_PREFIX_PREDEFINED

Predefined asset ids start with 0x02…

ASSET_ID_PREFIX_SBT_PREDEFINED

Predefined SBT asset ids start with 0x06…

ASSET_ID_PREFIX_SYNTHETIC

Synthetic asset ids start with 0x03…

CERES_ASSET_ID

DAI

DEFAULT_BALANCE_PRECISION

DEMETER_ASSET_ID

DOT

ETH

FIXED_PRECISION

HERMES_ASSET_ID

KARMA

KEN

KGOLD

KSM

KUSD

KXOR

PRUSD

PSWAP

SB

SBT_PRACS

SBT_PRCRDT

SBT_PRINVST

TBCD

TECH_ACCOUNT_MAGIC_PREFIX

This data is used as prefix in AccountId32, if it is representative for TechAccId encode twox hash (128 + 128 = 256 bit of AccountId32 for example).

USDT

VAL

VXOR

XOR

XST

XSTUSD

Traits

AssetInfoProvider

Trait to provide info about assets

AssetManager

AssetRegulator

Trait to manage permissions/regulations for assets operations

BuyBackHandler

Config

Common DEX trait. Used for DEX-related pallets.

DataFeed

DataFeed trait indicates that particular object could be used for querying oracle data.

DemeterFarming

DexInfoProvider

Trait to provide DEXInfo

EnabledSourcesManager

Implements trading pair EnabledSources storage

EnsureDEXManager

Check on origin that it is a DEX owner.

EnsureTradingPairExists

ExtendedAssetsManager

FromGenericPair

Abstract trait to get data type from generic pair name and data.

GetMarketInfo

Trait for bounding liquidity proxy associated type representing primary market in TBC.

GetPoolReserves

Trait for bounding liquidity proxy associated type representing secondary market.

IsRepresentable

IsRepresentation

IsValid

ItoaInteger

LiquidityProxyTrait

Indicates that particular object can be used to perform exchanges with aggregation capability.

LiquidityRegistry

LiquiditySource

Indicates that particular object can be used to perform exchanges.

LockedLiquiditySourcesManager

Implements trading pairs LockedLiquiditySources storage

OnNewSymbolsRelayed

Hook-like trait for oracles to capture newly relayed symbols.

OnPoolCreated

OnPoolReservesChanged

OnPswapBurned

General trait for passing pswap amount burned information to required pallets.

OnSymbolDisabled

OnValBurned

General trait for passing on the amount of burned VAL.

OrderBookManager

Trait to manage and interact with order book

PriceToolsProvider

PureOrWrapped

ReferencePriceProvider

ReferrerAccountProvider

SwapAction

Definition of a pending atomic swap action. It contains the following three phrases:

SwapRulesValidation

SyntheticInfoProvider

ToFeeAccount

ToOrderTechUnitFromDEXAndTradingPair

ToXykTechUnitFromDEXAndTradingPair

TradingPairSourceManager

Vesting

Trait to abstract interface of VestedRewards pallet, in order for pallets with rewards sources avoid having dependency issues.

WrappedRepr

XykPool

Functions

comm_merkle_op

Commutative merkle operation, is crypto safe, defined as hash(a,b) xor hash(b,a).

convert_block_number_to_timestamp

Converts block_number to timestamp

hash

hash_to_u128_pair

itoa

Converts integer into Vec string. Works with no-std

sort_with_hash_key

Sorting of keys and values by key with hash_key, useful for crypto sorting with commutative merkle operator.

with_transaction

Similar to #[transactional]

Type Definitions

AccountIdOf

Amount

AmountOf

Asset

Basic type representing asset.

AssetId32Code

This code is H256 like.

AssetIdOf

Balance

BalanceOf

BalancePrecision

BasisPoints

Type definition representing financial basis points (1bp is 0.01%)

CurrencyIdOf

DexIdOf

Fixed

Basic type representing assets quantity.

FixedInner

GetBaseAssetIdOf

Price