Module common::prelude

Re-exports

pub use super::Fixed;

pub use super::FixedInner;

pub use fixnum;

pub use super::serialization::*;

pub use super::weights::*;

Structs

AssetId32

This is wrapped structure, this is like H256 or Р512, extra PhantomData is added for typing reasons.

AssetName

AssetSymbol

BalanceUnit

BalanceUnit wraps Balance and provides proper math operations between divisible & non-divisible balances that have different precision.

BalanceWrapper

Wrapper type which extends Balance serialization, used for json in RPC’s.

BoundedString

ContentSource

CrowdloanTag

DEXInfo

Information about state of particular DEX.

Description

FixedWrapper

A convenient wrapper around Fixed type for safe math.

LiquiditySourceId

Identification of liquidity source.

LiquiditySourceTypeIter

An iterator over the variants of LiquiditySourceType

PswapRemintInfo

Rate

Information about received oracle symbol (price and last update time)

Scalar

BalanceUnit can be multiplied by scalars using this type.

SwapChunk

SwapOutcome

Amount of output tokens from either price request or actual exchange.

SymbolName

TradingPair

Trading pair data.

TryFromQuoteAmountError

TryFromSwapAmountError

TryFromSwapOutcomeError

Enums

AssetIdExtraAssetRecordArg

DEXId

DEX identifier.

FilterMode

InvokeRPCError

Common error which can arise while invoking particular RPC call in runtime.

LiquiditySourceType

Enumaration of all available liquidity sources.

ManagementMode

Oracle

List of available oracles

PredefinedAssetId

Asset identifier.

PriceVariant

QuoteAmount

RewardReason

Reason for particular reward during swap.

SwapAmount

Used to identify intention of caller either to transfer tokens based on exact input amount or exact output amount.

SwapVariant

Used to identify intention of caller to indicate desired input amount or desired output amount. Similar to SwapAmount, does not hold value in order to be used in external API.

TechAccountId

Enum encoding of technical account id, pure and wrapped records. Enum record WrappedRepr is wrapped represention of Pure variant of enum, this is useful then representation is known but backward mapping is not known.

TechAssetId

TechPurpose

Code of purpose for technical account.

Constants

ASSET_CONTENT_SOURCE_MAX_LENGTH

Max length of asset content source. The same value as IE URL length. It should enough for any URI / IPFS address (CID)

ASSET_DESCRIPTION_MAX_LENGTH

Max length of asset description, it should be enough to describe everything the user wants

CERES_ASSET_ID

DAI

DEFAULT_BALANCE_PRECISION

DEMETER_ASSET_ID

DOT

ETH

HERMES_ASSET_ID

KSM

KUSD

PSWAP

TBCD

USDT

VAL

XOR

XST

XSTUSD

Traits

AssetInfoProvider

Trait to provide info about assets

BuyBackHandler

Config

Common DEX trait. Used for DEX-related pallets.

DataFeed

DataFeed trait indicates that particular object could be used for querying oracle data.

DemeterFarming

DexInfoProvider

Trait to provide DEXInfo

EnabledSourcesManager

Implements trading pair EnabledSources storage

EnsureDEXManager

Check on origin that it is a DEX owner.

EnsureTradingPairExists

FromGenericPair

Abstract trait to get data type from generic pair name and data.

GetMarketInfo

Trait for bounding liquidity proxy associated type representing primary market in TBC.

GetPoolReserves

Trait for bounding liquidity proxy associated type representing secondary market.

IsRepresentable

IsRepresentation

IsValid

LiquidityProxyTrait

Indicates that particular object can be used to perform exchanges with aggregation capability.

LiquidityRegistry

LiquiditySource

Indicates that particular object can be used to perform exchanges.

LockedLiquiditySourcesManager

Implements trading pairs LockedLiquiditySources storage

OnNewSymbolsRelayed

Hook-like trait for oracles to capture newly relayed symbols.

OnPoolCreated

OnPoolReservesChanged

OnPswapBurned

General trait for passing pswap amount burned information to required pallets.

OnSymbolDisabled

OnValBurned

General trait for passing on the amount of burned VAL.

PriceToolsProvider

PureOrWrapped

ReferencePriceProvider

ReferrerAccountProvider

SwapAction

Definition of a pending atomic swap action. It contains the following three phrases:

SwapRulesValidation

SyntheticInfoProvider

ToFeeAccount

ToOrderTechUnitFromDEXAndTradingPair

ToXykTechUnitFromDEXAndTradingPair

TradingPairSourceManager

Vesting

Trait to abstract interface of VestedRewards pallet, in order for pallets with rewards sources avoid having dependency issues.

WrappedRepr

XykPool

Type Definitions

AccountIdOf

AssetId32Code

This code is H256 like.

Balance

BalancePrecision

DexIdOf